Add iterator for yielding all token details not just int rank
This was pretty easy to do but the `encode` (as opposed to
`encode_ordinary`) version which knows about special tokens and encodes
them separately will be more of a challenge.  I doubt it's actually
worth this hassle.  If it's a big deal to implement streaming
tokenizetion, the right way to do it would be to chunk the incoming text
anyway.  My implementation already wouldn't work on an iterator of
string chunks; it needs one string as input which it then tokenizes.