Refactor how messages are submitted
This actually makes the code more crufty but it's a step along the way.  Now the React code doesn't just assume the message got sent; it waits until there's a response.  This is shitty UX which I'll fix next, but it will be less confusing in the error case when the message wasn't actually able to be sent.

Up next: refactor this more to use events from the backend when messages are pushed to the session but not yet sent, when they fail to send or are successfully sent, and when a new assistant message is received.