Improve error handling and diagnostic features
This enables logging from Rust and Typescript that will all go to stdout and to the webview console.  It also switches from stringly-typed errors to something more structured.  Errors sending messages are now handled in a more graceful way.

Still not yet done is the ability to retry sending a message that previously failed due to an error.